+++ gfx/resources/OtFrameBuffer.cpp
@@ -95,11 +95,23 @@ void OtFrameBuffer::update(int w, int h) {
 		uint64_t blit = blitTarget ? BGFX_TEXTURE_BLIT_DST : 0;
 
 		if (colorTextureType != OtTexture::noTexture) {
-			colorTexture = bgfx::createTexture2D(w, h, false, 1, (bgfx::TextureFormat::Enum) colorTextureType, flags | blit);
+			colorTexture = bgfx::createTexture2D(
+				static_cast<uint16_t>(w),
+				static_cast<uint16_t>(h),
+				false,
+				1,
+				(bgfx::TextureFormat::Enum) colorTextureType,
+				flags | blit);
 		}
 
 		if (depthTextureType != OtTexture::noTexture) {
-			depthTexture = bgfx::createTexture2D(w, h, false, 1, (bgfx::TextureFormat::Enum) depthTextureType, flags | blit);
+			depthTexture = bgfx::createTexture2D(
+				static_cast<uint16_t>(w),
+				static_cast<uint16_t>(h),
+				false,
+				1,
+				(bgfx::TextureFormat::Enum) depthTextureType,
+				flags | blit);
 		}
 
 		// create framebuffer
